Add a control fd to the port
Similar to the alcove child processes, generate an event when the alcove
port process has called exec() by using a file descriptor opened with
the close-on-exec flag.

The standard way of communicating between two unrelated Unix processes
(beam and the alcove port) is by using a fifo. The fifo is opened
read/write in the alcove port port to prevent open(2) from blocking.

The Erlang side opens the fifo in read-only mode using another Erlang
port as described here:

https://gist.github.com/jaredmorrow/1c342c6e9156eddd20b2

If the alcove port calls exec(), the control fd will be closed and the
Erlang port associated with the fifo will be exit.