Support for re-exec'ing the port process
Each process in the fork chain is a recursive call. The stack for each
successive child process will grow until eventually the stack overflows
and the process is killed with a segfault (which is why the fork depth
is limited to 16 processes by default).

Since each child process is a copy of the parent, some security
features like ASLR will be ineffective.

Having the port exec() itself and continue in call mode fixes these
issues. To do this, the port:

* uses an internal environment variable (ALCOVE_SHLVL) to indicate
  whether the port is being exec'ed by beam or by itself

* sets up the control fd's if exec'ed by beam

To exec the port, the FD_CLOEXEC flag on the control fd's (3: signal read,
4: signal write, 5: control on exec fd) must first be unset using
fcntl/5.

The port does not exec() by default because:

* the alcove binary may not be accessible

  The parent may have called chroot() or other call that changes the
  filesystem namespace.

* the process may not have permission to read or exec the binary

  The parent may have changed to another uid.