Initialize fd's based on a command line argument
Use a new command line switch (-d) instead of using an environment
variable to indicate whether the control fd's should be opened by
main(). "-d" takes an integer argument specifying the fork depth (the
port is 0, the first child is 1, etc).

The setenv() was causing processes on some systems (OpenBSD, Solaris) to
crash during the test run. On Solaris, an exit value of {error,enomem}
would sometimes be seen.