Perform bounds checks on external term format
Check the claimed length of the term encoded in the external term format
will fit into the buffer.

Move the first pass check of the cstruct list to use the alcove_decode_*
functions. This change is not really necessary: the call to
alcove_get_type() will error out before the calls to ei.