cstruct: limit allocated size to MAXMSGLEN
Crash early if the total number of bytes allocated in the cstruct will
not fit in 2 bytes. For example, passing in the following list will now
return badarg (crashing the calling erlang process) instead of forcing the
Unix process to exit when the results cannot be encoded in the response:

    alcove:alloc(Drv, [Child], [{ptr, 40000}, {ptr, 40000}]).

The number of allocated bytes may be smaller than the number of bytes
written to the buffer so it is critical that both values are checked.

It is still possible for the caller to unintentionally pass in values
which can't be represnted in the wire format, crashing the Unix process.

Also limit the number of fields in the cstruct to 1023. The value was
taken from:

http://stackoverflow.com/a/19938035

The comment was copied from ISO/IEC 9899:TC3
(Committee Draft Septermber 7, 2007 WG14/N1256):

http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf