alcove_event_loop: unused parameter warnings
Good discussion on the different approaches to silence the unused
parameter warnings:

http://stackoverflow.com/questions/1486904/how-do-i-best-silence-a-warning-about-unused-variables

Using __attribute__((unused)) will work with both gcc and clang but looks
a bit weird with pointers in the function signature. The advantage is that
will gcc will error if the variable is accidentally used in the future.

"(void)var" is portable and, at least for function pointers, shows the
intent. According to stack overflow, some levels of optimization may
cause it to be removed.