Move from eunit to common test
Create a separate test suite for testing the port message format.

Clean up the tests so that the state of one test does not rely on
changes in a previous test, e.g., the previous test doing a chroot.

Fork a new alcove port for each test. The advantage is that tests can
run independently and messages do not accumulate in the mailbox. The
downside is that a single port process is not being exhaustively tested
which might turn up memory leaks, stack overflows, ...