Allow caller to handle SIGCHLD
Modify handling of SIGCHLD in the event loop so the caller can handle
reaping the child processes.

Introduce waitpid/4 to allow the parent to wait for children. The return
value includes the status but should probably be extended to include the
exit status and/or signal.

Make the handling of signals more uniform. For SIGCHLD:

    SIG_DFL:   event loop waits for child, closes file descrptors
    SIG_IGN:   event loop ignores child exit. Caller must handle child.
    SIG_CATCH: event loop sends event to caller. Caller must handle
               child.

This change allows experimenting with systems calls that reparent the
child process like ptrace(2).