alcove: modify handling of bad message headers
If an invalid message header was received, the process would call exit
and write to stderr:

    process 1 (bad hdr) -> process 2 (process exits) -> process 3
    process 1 <- process 2 (process exits) <- process 3 (bad hdr)

Avoid writing to stderr and treat the error in the same way as a write
error:

    process 1 (bad hdr) -> process 2 (write error, child stdout closed) -> process 3 (process exits)
    process 1 <- process 2 (write error, stdout closed) <- process 3 (bad hdr)

The semantics are similar but still not entirely correct. Possibly the
process should return 'badmsg' or 'badhdr' and force an exception.