alcove: ensure order of messages on child exit
When a process quickly forks, exec()'s and exits there is a race
condition between the parent receiving SIGCHLD and poll() returning
ready for the child fd's.

On linux, the order always appears to be:

    fdctl_closed
    <stdout>
    exit_status

On freebsd and openbsd:

    exit_status
    fdctl_closed
    <stdout>

Partially fix by having the parent first poll the child's fd's in the
SIGCHLD handler. This is just a partial fix because:

* prx:sh/2 and prx:cmd/2 now work on freebsd with occasional failures

* alcove_SUITE:stream test now fails on freebsd

* prx_SUITE:eof test still fails on freebsd and openbsd