sighandler: restrict to async signal safe funcs
The usage of static allocation and memcpy() in a signal handler is not
stricly correct. Make the signal handlers uniform by always requesting
the signal information and write it to the process via a pipe. Since
the write is atomic, a call to the handler function by another signal
will not interfere with the write: either the write will complete or,
possibly, write will return -1 (EINTR) and the process will crash.

The function pointers to the signal handlers are used to distinguish
between the default behaviour (for SIGCHLD handling) and returning a
3-tuple signal with the siginfo_t. This requires calling sigaction(2) so
the base allowed seccomp policy must be extended to include
SYS_RT_SIGACTION.