rebar.config.script: write to TMPDIR
The tests for seccomp and setns(2) would fail if alcove was a dependency
for a project without a "test" directory.

Use the TMPDIR environment variable to specify where to write the test
programs:

    TMPDIR=./_build rebar3 do clean, compile, ct

If TMPDIR is unset, the default is /tmp. The files are created O_EXCL
to prevent race conditions.

Also try to factor out some of the boilerplate as anonymous functions. The
script can be tested by doing:

    file:script("rebar.config.script", [{'CONFIG', ""}]),
    os:getenv("ALCOVE_DEFINE").