alcove_drv: use tagged tuples for errors
Return a tagged tuple to indicate an error rather than erroring out in
the alcove_drv interfaces. In either case, the crash occurs in the
context of the caller, not in the driver process.

The port process can return any term. In general, errors are returned as
bare atoms ... except for 'ok', 'true' and maybe a few others.