alcove_drv: try to be explicit about error messages
Begin specifying the error values from the port process.

The port has a few ways of returning errors:

* as a response to a call: an invalid argument or call is not supported
  on this platform

* message was sent to a Unix process that does not exist or has been
  terminated

* a process was sent a signal

In these cases, the owning erlang process is killed.

With this change, control messages, except for fdctl_closed, will also
result in an error. Probably a future change should leave these in the
caller's mailbox.