stdin: non-blocking write to the child process
Mark the file descriptor for stdin from the parent to child as
non-blocking.

With blocking writes, it is possible for the parent process to block if
the child is not reading stdin, eventually causing the port process to
block.

A write that returns EAGAIN will send a "badwrite" message to the owner
of the port. This should cause the owner to crash but the exception
may be caught and the write re-tried:

* if the amount of data is below the atomic write limit, it can be safely
  retried (minimum of 512 bytes; Linux has a limit of 4096 bytes).

* if the amount of data is greater than the atomic write limit, a
  partial write may have occurred before the error.

See pipe(7) for details on PIPE_BUF handling.