select: fix for 32-bit platforms
The optional timeout for select(2) is a long integer. The internal
alcove protocol passes the value as a long long or 64-bit integer. The
value was (incorrectly) written to the struct using a cast.

Use a temporary 64-bit value to hold sec/usec and truncate the integer
on 32-bit platforms.