Replace sysconf(3) with getrlimit(RLIMIT_NOFILE)
Use the process resource limits to set the alcove runtime limits. The
idea is to have most of the alcove command line options replaced by
calculating them based on the resource limits.

Further changes are required to calculate the number of children based
on the RLIMIT_NOFILE. In particular, the child array is not re-sized and
so can be overflowed if the RLIMIT_NOFILE is increased.