ioctl: use int64_t for ioctl request arg
Traditionally ioctl(2) has used a signed 32-bit integer for the request
argument. The ioctl macros are made for 32-bit values. At some point,
request was changed from an int32_t to an unsigned long, either a
uint32_t or uint64_t depending on the platform. See this discussion:

https://sourceware.org/bugzilla/show_bug.cgi?id=14362

On FreeBSD, the values for some ioctl requests are greater than
INT32_MAX causing ioctl to crash with badarg:

    {error, enotcapable} = alcove:ioctl(Drv, [Child], FD, tiocmset, <<>>)

Use a signed 64-bit integer for ioctl requests. Since requests are
internally unsigned 32-bit values, the range should be sufficient. It is
possible that some requests may be greater than a signed 64-bit value but
these would not be representable on a 32-bit system.

To be correct, alcove should probably add uconstant and uconstant64.