alcove_drv: match utf atom from port
The ei library distributed with erlang 20 may encode the atom "ok"
returned on port initialization as a utf atom, causing the pattern
match to fail and the port to block in receive forever. Mysteriously,
when called from prx, the same code returns a latin1 encoded atom.

The case statement is probably not necessary: if at any point before
'ok' is returned there is an error, the port will crash and exit. No
errors are returned. If the binary sent by the term is corrupted,
alcove_codec:decode/1 will crash, forcing the port to exit.