Send signal to child processes on EOF
If the stdin of an alcove process is closed, send a SIGTERM to any
exec'ed child processes by default.

Usually an EOF on stdin is caused by the erlang VM exiting though it could
also be caused by the parent of an alcove process exiting, e.g., from
a signal.

For example:

    beam -+- alcove1 -+- cat
                     `- sleep
          |- alcove2 --- nc -k -l 8080

* beam exits

* stdin of alcove1 is closed, alcove1 exits

* stdin of cat is closed, cat exits

* stdin of sleep exits: sleep does not monitor stdin and so does not
  exit

* stdin of alcove2 is closed, alcove2 exits

* stdin/stdout of nc is closed: nc does not monitor stdin and does not exit

  If/when nc writes to stdout, it will be killed by SIGPIPE.

The signal can be set by using alcove:setopt/2:

    alcove:setopt(Drv, [], signaloneof, 9)

`signaloneof` mimics the behaviour of bash's huponexit and lxc-stop's
lxc.haltsignal. bash sends a SIGHUP and lxc sends a SIGTERM by
default, followed by a SIGKILL after 60 seconds.