alcove: fix dialyer warning
Remove match of the alcove_pipe tuple from stdin/3. An alcove_pipe error
is caused by a write to a pipe with a full buffer (EPIPE), e,g., a write
to a subprocess that is not reading input from stdin.

stdin/3 is a cast.  Errors will be returned via a message sent to the
caller's mailbox (or by calling stdout/2,3 or stderr/2,3).

Alternatively, stdin could be converted to a synchronous call, either by
returning a value after a write in the port. Another option might be
to check for alcove_pipe before successive calls to stdin.