setopt: match behaviour of setcpid/5
The flowcontrol setting in setopt/4 was a boolean enabling or disabling
the default flowcontrol for a new process. Enabled (1) meant no stdio
would be read from the subprocess similar to {active,false} in gen_tcp;
disabled (0) meant no flow control was enforced similar to
{active,true}.

setcpid/5 uses a count: 0 == no further stdio read, -1 == no flow
control, >0 == this many messages read

The difference was confusing:

    setopt == 0, no flow control
    setcpid == 0, flow controlled (no further output)

The behaviour of setopt is now the same as setcpid. Setting a
flowcontrol value of (int32_t)-1 disables flow control. Flow control is
disabled by default.

The flowcontrol test is now failing on linux. Previously it would fail
sporadically and would consistently fail on OpenBSD and FreeBSD. It
seems after calling exec() the stdio fd's are added to the poll set,
allowing one message from the subprocess to be sent. Further messages
are flowcontrolled properly.