flowcontrol: disable if subprocesses has exited
A child process that called exec() then exited could cause the parent
process to block or loop in poll() if flow control was enabled:

* poll() returns POLLHUP (child end of pipe closed) for the child
  stdout fd

  read() would return EOF but the read is skipped due to flowcontrol.

  In this case, the parent process will quickly poll(), return then
  re-enter poll.

* poll() returns POLLNVAL (for mysterious reasons: possibly alcove has
  already closed its end of the pipe)

  read() would return an error but again is skipped due to flowcontrol.

  The symptom is a subsequent call will cause the port to block in
  read().

Disable the flowcontrol checks if the process has exited (SIGCHLD was
received for the subprocess).

Currently errors in the poll() revents field are only used to indicate
a read() is pending. Errors are returned by doing a read() from the
file descriptor. An improvement might be to use the revents errors and
avoid the read.