alcove: fix hang when subprocess exits
If alcove forks a process tree and the immediate subprocess is
terminated, the child subprocesses of the subprocess are re-parented to
init(1):

```
1:
alcove -- sh -- sleep inf
          ^ SIGTERM

2:
alcove
init -- sleep inf
```

alcove reads the stdout of the `sh` process. Usually the read returns 0
bytes (EOF), the alcove process closes the pipe between the
processes and alcove resumes processing any commands.

In the case with orphaned subproceses, the behaviour is surprising:
the read blocks until all subprocesses are terminated.

Fix by temporarily making the pipe file descriptor non-blocking before
performing the read.