pledge: replace usage of null with NIL list
The atom 'null' or 'NULL' was used to signify the NULL value for
pledge(2). An empty string was equivalent to "no promises".

    alcove:pledge(Drv, [Pid], "stdio exec", "")

A promise string is iodata. The above is equivalent to:

    alcove:pledge(Drv, [Pid], [<<" stdio ">>, <<" exec ">>], "")

An empty list (NIL list) could represent either "no promises" or the
NULL value:

    % except [] == ""
    alcove:pledge(Drv, [Pid], [], "")

    % or
    alcove:pledge(Drv, [Pid], <<"stdio exec">>, "")
    alcove:pledge(Drv, [Pid], <<"">>, "")

    % changed to either:
    alcove:pledge(Drv, [Pid], [], <<"">>)
    alcove:pledge(Drv, [Pid], [], <<>>)

Change the pledge arguments to:

* use [] as NULL to be consistent with other functions
* in the examples, use binaries for arguments