code generation: include static code, edoc
Include static code as a header instead of storing it as a string inside
an escript. If source line numbers do not match up with the edoc, the
header could be read directly into alcove.erl by the escript.

Generate documentation so it is available to edoc and the shell.

The compiler has some directives that could have been used instead of
embedding the docs and specs:

* headerfile

    * warns if doc and spec are included
    * typer doesn't expand the included spec

* docfile: works, downside is code/docs are in separate files