filter/1,3,4: fix and simplify
Simplify filtering by passing a bit set to the port and OR'ing each
byte. Removes the ambiguity in the erl external format between lists and
strings and some buggy code: the second check_filter() invocation used the
pointer to the first call list but the length of the second.

Fix the handling of nil lists. The term was not read so if the first
call list was [], the second call list would be ignored. Note: an empty
binary (<<>>) now has the same effect as the nil list.

Update filter/1,3,4 with examples.