setrlimit: use /dev/fd for open fd's
The behaviour of calling setrlimit(RLIMIT_NOFILE) with a value below
the highest open file descriptor differs by platform:

* FreeBSD: returns an error
* Linux, OpenBSD, Solaris: allowed

This behaviour is useful for restricting the process:

* open fd's
* set RLIMIT_NOFILE to 0
* new file descriptors can't be opened

alcove uses the value of RLIMIT_NOFILE though to calculate the maximum
number of child processes. If RLIMIT_NOFILE is below the current highest
fd, the subprocess will not be addressable. To avoid that situation,
calls to setrlimit/4,5 check for open fd's in the range between the
requested and current rlimit.

There are a few issues:

* when checking for open fd's, the code checked the same fd repeatedly,
  not the incremented fd

* additionally, the check occurred for the fd after the limit, i.e.,
  RLIMIT_NOFILE=64 is 0-63, the check allowed 0-64

* if RLIMIT_NOFILE is set to a large number (e.g., crostini sets the
  default to 1048576), calling setrlimit/4,5 is slow

Fix by:

* iterating through /dev/fd for open file descriptors
* falling back to calling fcntl(2) if the directory is not accessible
* using /dev/fd for portability: on Linux, it is a symlink to
  /proc/self/fd