maxchild: (re)decouple limit from RLIMIT_NOFILE
The limit on the number of child processes was removed in a series of
commits (6f12cde..465ad5f6 in 2017(!)). The maxchild limit was derived
from the current value of RLIMIT_NOFILE.

Even with a low setting for file descriptors, the number of allowed
child processes was high: (8192-6)/6 = 1364

With a high number of file descriptors, performance is slow as alcove
does a linear scan of the fd set.

For example, on crostini, the default fdlimit is 1048576 or 174761
subprocesses:

* 174761 processes exceeds the default RLIMIT_NPROC value of 26479

* results in a unsigned integer overflow since the type of an fdset is
  a uint16_t

* a typical process supervisor (init or inetd style) may set a high
  fdlimit for services

Reinstate the maxchild setting with a reduced default value (64) since
a common case is for the control processes to fork a single process
calling exec.

The behaviour of maxchild:

* the value can now be set at runtime
* the value cannot be lower than the current number of child processes
* the number of file descriptors required for the value cannot be higher
  than the process RLIMIT_NOFILE
* lowering setrlimit(RLIMIT_NOFILE) will error as above or lower the
  maxchild setting
* raising setrlimit(RLIMIT_NOFILE) has no effect on the maxchild setting
* setting maxchild to 0 prevents the control process from fork()'ing
  (maxforkdepth has the same effect)