0.40.1: cstruct: fix max allocated check
A cstruct has 2 lengths: the number of serialized bytes for the struct
to be represented as a binary and the amount of memory allocated.

    # 64-bit pointer: 8 bytes
    # memory required for value: 10 bytes
    [{ptr, "0123456789"}]

Commit cb31b4d165858322 limited the total amount of memory allocated by a
cstruct (according to the commit message) to a 16-bit value. The actual
check tested against the passed in buffer length.

For most users of `alcove_decode_cstruct()` this statement was correct
since the passed in length of the buffer is MAXMSGLEN (UINT16_MAX).

`jail/3` uses a value of `sizeof(struct jail)` (56 bytes with padding)
which contains fields of type `char *`. If the caller, e.g., sets a
hostname or jailname > 7 bytes (excluding trailing NULL), the operation
would fail with badarg.

The idea was to protect a program from crashing alcove with an OOM by
accidentally passing in a very large value. The caller will instead get
a badarg exception.

Limit the total allocated memory by a cstruct to UINT16_MAX.